frogger animation

+++ Sprite.java
@@ -20,8 +20,8 @@ public abstract class Sprite {
     protected byte positionY            = 0;
     protected short pixelPosX           = 0;
     protected short pixelPosY           = 0;
-    protected short inBetweenX          = 0;
-    protected short inBetweenY          = 0;
+    protected double inBetweenX         = 0;
+    protected double inBetweenY         = 0;
     protected byte width                = 0;
     protected byte height               = 0;
     protected byte offsetTop            = 0;
